feat(dashmate): integrate DMT to setup
